Consolidated ShapesFactoryUtils into ShapesFactory. Fixed bug that kept icons from being dimmed when a node or connection was hovered/highlighted.  Removed commented out code.  Added a very detailed todo message in ShapeParent.
